requirements.txt
flask==3.0.0
flask-cors==4.0.0
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
python-dateutil==2.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
